+++ backend/src/services/resolution.rs
@@ -29,14 +29,14 @@ pub async fn resolve_reference(
           sv.version             AS version,
           sv.code                AS code,
           sv.dependencies        AS dependencies,
-          ascore.formula_version AS quality_formula_version,
-          ascore.freshness       AS quality_freshness,
-          ascore.adoption        AS quality_adoption,
-          ascore.reliability     AS quality_reliability,
-          ascore.abandonment     AS quality_abandonment,
-          ascore.overall         AS quality_overall,
-          ascore.flags           AS quality_flags,
-          ascore.computed_at     AS quality_computed_at
+          ascore.formula_version      AS quality_formula_version,
+          ascore.freshness::float8    AS quality_freshness,
+          ascore.adoption::float8     AS quality_adoption,
+          ascore.reliability::float8  AS quality_reliability,
+          ascore.abandonment::float8  AS quality_abandonment,
+          ascore.overall::float8      AS quality_overall,
+          ascore.flags                AS quality_flags,
+          ascore.computed_at          AS quality_computed_at
         FROM libraries l
         JOIN snippets s          ON s.library_id = l.id
         JOIN snippet_versions sv ON sv.snippet_id = s.id
